requirements.txt
Flask==2.2.5
Adyen==13.4.0
Flask-Mail
python-dotenv
